requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot>=2.0.0
PyMySQL>=1.0.0
DBUtils>=3.0.0
PyYAML>=6.0
aiohttp>=3.8.0
requests>=2.28.0